FastAPI-MongoDB/requirements.txt
fastapi==0.128.6
passlib==1.7.4
pydantic==2.12.5
pymongo==3.12.0
pymongo[srv]
python_jose==3.5.0
python-dotenv
uvicorn
python-multipart
pathlib
cloudinary
